requirements.txt
cffi==2.0.0
cryptography==48.0.0
greenlet==3.5.0
Markdown==3.10.2
mysql-connector-python==9.7.0
pycparser==3.0
PyMySQL==1.1.3
python-dotenv==1.2.2
python-frontmatter==1.1.0
PyYAML==6.0.3
SQLAlchemy==2.0.49
typing_extensions==4.15.0
